django/django_orm_example_23/requirements.txt
alabaster==0.7.8
amqp==1.4.9
anyjson==0.3.3
astroid==1.4.4
Babel==2.3.4
behave==1.2.5
billiard==3.3.0.23
celery==3.1.23
colorama==0.3.6
Django==1.9.5
django-behave==0.1.5
django-celery==3.1.17
django-debug-toolbar==1.4
django-querycount==0.4.1
django-split-settings==0.1.3
djangorestframework==3.3.3
docopt==0.6.2
docutils==0.12
enum34==1.1.4
gitdb==0.6.4
GitPython==1.0.2
imagesize==0.7.1
Jinja2==2.8
joe==0.0.7
kombu==3.0.35
lazy-object-proxy==1.2.1
Markdown==2.6.6
MarkupSafe==0.23
mercurial==3.6.3
parse==1.6.6
parse-type==0.3.4
pbr==1.8.1
pep8==1.7.0
Pillow==3.2.0
powerline-status==2.3
Pygments==2.1.3
pylint==1.5.4
pytz==2016.3
redis==2.10.5
selenium==2.53.2
six==1.10.0
smmap==0.9.0
snowballstemmer==1.2.1
Sphinx==1.4.1
sqlparse==0.1.19
stevedore==1.10.0
virtualenv==14.0.5
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.26.0
wrapt==1.10.6
